pyproject.toml
[tool.poetry]
name = "django-keycloak"
version = "0.2.9"
description = "Integrate Keycloak with Django 4.2+"
homepage = 'https://github.com/Tehnari/django-keycloak'
repository = 'https://github.com/Tehnari/django-keycloak'
keywords = ["django", "keycloak", "auth", "django42", "python311","django-keycloak"]
license = "MIT"
authors = [
  "Constantin Sclifos <sclifcon@gmail.com>",
  "Samuel C. Tyler <skamansam@gmail.com>",
  "Ahmad Dabo <dabo.cs@gmail.com>",
  "Peter Slump <peter@yarf.nl>",
]
readme = "README.rst"
packages = [{include = "django_keycloak"}]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Tehnari/django-keycloak/issues"

[tool.poe.tasks]
test         = "pytest --cov=my_app"                         # a simple command task
serve.script = "my_app.service:run(debug=True)"              # python script based task
tunnel.shell = "ssh -N -L 0.0.0.0:8080:$PROD:8080 $PROD &"   # (posix) shell based task


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Django = ">=4.2"
python-keycloak = ">=4.1,<5"
httpx = "<0.28"

[tool.poetry.group.dev.dependencies]
django-dynamic-fixtures = "^0.2.1"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
mock = "^5.1.0"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
